refactor: small changes
